Use the PFX option and rename to AdCs
