+++ lib/msf/core/exploit/remote/cert_request.rb
@@ -34,15 +34,22 @@ module Msf
         application_policies: application_policies
       )
 
+      pkcs12 = nil
+      if opts.key?(:pkcs12)
+        pkcs12 = opts[:pkcs12]
+      elsif datastore['PFX'].present?
+        pkcs12 = OpenSSL::PKCS12.new(File.binread(datastore['PFX']))
+      end
+
       on_behalf_of = opts[:on_behalf_of] || (datastore['ON_BEHALF_OF'].blank? ? nil : datastore['ON_BEHALF_OF'])
       status_msg << " - on behalf of: #{on_behalf_of}" if on_behalf_of
-      if @pkcs12 && on_behalf_of
+      if pkcs12 && on_behalf_of
         vprint_status("Building certificate request on behalf of #{on_behalf_of}")
         csr = Rex::Proto::X509::Request.build_on_behalf_of(
           csr: csr,
           on_behalf_of: on_behalf_of,
-          cert: @pkcs12.certificate,
-          key: @pkcs12.key,
+          cert: pkcs12.certificate,
+          key: pkcs12.key,
           algorithm: algorithm
         )
       end
