+++ lib/msf/core/exploit/remote/http/web_enrollment.rb
@@ -9,7 +9,7 @@ module Msf
         module WebEnrollment
 
           include Msf::Exploit::Remote::CertRequest
-          include Msf::Exploit::Remote::LDAP::ActiveDirectory::AdcsOpts
+          include Msf::Exploit::Remote::LDAP::ActiveDirectory::AdCsOpts
 
           def get_cert_templates(authenticated_client)
             print_status('Retrieving available template list, this may take a few minutes')
@@ -117,7 +117,7 @@ module Msf
                                      pkcs12.to_der,
                                      'certificate.pfx',
                                      info)
-            print_good("Certificate for #{connection_identity} using template #{cert_template} saved to #{stored_path}")
+            print_good("Certificate using template #{cert_template} saved to #{stored_path}")
             certificate
           end
         end
