--- lib/msf/core/exploit/remote/ldap/active_directory/adcs_opts.rb
@@ -1,30 +1,0 @@
-module Msf
-  ###
-  #
-  # This module exposes methods for querying a remote LDAP service
-  #
-  ###
-  module Exploit::Remote::LDAP::ActiveDirectory
-
-    module AdcsOpts
-
-
-    def initialize(info = {})
-      super
-
-      register_options([
-                         OptString.new('ALT_DNS', [ false, 'Alternative certificate DNS' ]),
-                         OptString.new('ALT_SID', [ false, 'Alternative object SID' ]),
-                         OptString.new('ALT_UPN', [ false, 'Alternative certificate UPN (format: USER@DOMAIN)' ]),
-                         OptString.new('CERT_TEMPLATE', [ true, 'The certificate template', 'User' ]),
-                         OptPath.new('PFX', [ false, 'Certificate to request on behalf of' ]),
-                         OptString.new('ON_BEHALF_OF', [ false, 'Username to request on behalf of (format: DOMAIN\\USER)' ]),
-                       ])
-      register_advanced_options([
-                                  OptEnum.new('DigestAlgorithm', [ true, 'The digest algorithm to use', 'SHA256', %w[SHA1 SHA256] ]),
-                                  OptEnum.new('RSAKeySize', [ true, 'RSA key size in bits for CSR generation', '2048', %w[1024 2048 3072 4096 8192] ])
-                                ])
-    end
-    end
-  end
-end
