+++ lib/msf/core/exploit/remote/ms_icpr.rb
@@ -15,7 +15,8 @@ module Exploit::Remote::MsIcpr
 
   include Msf::Exploit::Remote::SMB::Client::Ipc
   include Msf::Exploit::Remote::DCERPC
-  include Msf::Exploit::Remote::LDAP::ActiveDirectory::AdcsOpts
+  include Msf::Exploit::Remote::CertRequest
+  include Msf::Exploit::Remote::LDAP::ActiveDirectory::AdCsOpts
 
   # [2.2.2.7.7.4 szOID_NTDS_CA_SECURITY_EXT](https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-wcce/e563cff8-1af6-4e6f-a655-7571ca482e71)
   OID_NTDS_CA_SECURITY_EXT = '1.3.6.1.4.1.311.25.2'.freeze
@@ -48,35 +49,6 @@ module Exploit::Remote::MsIcpr
       Opt::RPORT(445)
     ], Msf::Exploit::Remote::MsIcpr)
 
-  end
-
-  def setup
-    errors = {}
-    if datastore['ALT_SID'].present? && datastore['ALT_SID'] !~ /^S(-\d+)+$/
-      errors['ALT_SID'] = 'Must be a valid SID.'
-    end
-
-    if datastore['ALT_UPN'].present? && datastore['ALT_UPN'] !~ /^\S+@[^\s\\]+$/
-      errors['ALT_UPN'] = 'Must be in the format USER@DOMAIN.'
-    end
-
-    if datastore['ON_BEHALF_OF'].present?
-      errors['ON_BEHALF_OF'] = 'Must be in the format DOMAIN\\USER.' unless datastore['ON_BEHALF_OF'] =~ /^[^\s@]+\\\S+$/
-      errors['PFX'] = 'A PFX file is required when ON_BEHALF_OF is specified.' if datastore['PFX'].blank?
-    end
-
-    @pkcs12 = nil
-    if datastore['PFX'].present?
-      begin
-        @pkcs12 = OpenSSL::PKCS12.new(File.binread(datastore['PFX']))
-      rescue StandardError => e
-        errors['PFX'] = "Failed to load the PFX file (#{e})"
-      end
-    end
-
-    raise OptionValidateError, errors unless errors.empty?
-
-    super
   end
 
   def request_certificate(opts = {})
